packages/mp/pyproject.toml
[project]
name = "mp"
version = "1.29.3"
description = "General Purpose CLI tool for Google SecOps Marketplace"
readme = "README.md"
authors = [
    { name = "Tal Shapir", email = "talshapir@google.com" },
    { name = "Iddo Arie", email = "iddoa@google.com" },
    { name = "Dori Aviram", email = "doriaviram@google.com" },
    { name = "Amit Joseph", email = "amitjoseph@google.com" },
    { name = "Hagit Eliahu", email = "haggit@google.com" },
]
requires-python = "==3.11.*"
keywords = [
    "mp",
    "google",
    "secops",
    "soar",
    "marketplace",
    "content hub",
]

classifiers = [
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3 :: Only",
]
dependencies = [
    "libcst>=1.8.2",
    "pip>=25.1.1",
    "Pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "ruff>=0.15.12",
    "defusedxml>=0.7.1",
    "toml>=0.10.2",
    "typer>=0.16.0",
    "uv>=0.11.8",
    "requests>=2.32.4",
    "jinja2>=3.1.6",
    "deepdiff>=8.6.1",
    "platformdirs>=4.5.0",
    "ty>=0.0.34",
    "ruamel-yaml == 0.17.40",
    "google-genai>=1.74.0",
    "tenacity>=9.1.2",
    "anyio>=4.12.1",
    "toon-format",
    "packaging>=26.0",
    "questionary>=2.1.1",
]

[dependency-groups]
dev = [
    "hypothesis>=6.147.0",
    "pyrefly>=0.63.1",
    "pytest>=9.0.3",
    "pytest-cov>=6.2.1",
    "pytest-xdist>=3.7.0",
    "types-defusedxml>=0.7.0.20260408",
    "types-pyyaml>=6.0.12.20260408",
    "types-requests>=2.33.0.20260408",
    "types-toml>=0.10.8.20260408",
]

[project.scripts]
mp = "mp.main:main"
wmp = "mp.main:main"

[build-system]
requires = ["uv_build>=0.11.3,<0.12"]
build-backend = "uv_build"

[project.urls]
Repository = "https://github.com/chronicle/content-hub/tree/main/packages/mp"

[tool.pytest.ini_options]
pythonpath = "src"
addopts = [
    "--import-mode=importlib",
]
testpaths = [
    "tests/",
]

[[tool.uv.index]]
url = "https://pypi.org/simple"
default = true

[tool.uv.sources]
toon-format = { git = "https://github.com/toon-format/toon-python.git" }
